+++ frontend/src/components/home/measurements/sections/ClientSection.tsx
