+++ frontend/src/components/home/measurements/sections/InvoiceSection.tsx
@@ -1,4 +1,3 @@
-import React from 'react'
 
 const InvoiceSection = () => {
   return (
