+++ frontend/src/components/home/measurements/sections/ProjectSection.tsx
@@ -1,4 +1,3 @@
-import React from 'react'
 
 const ProjectSection = () => {
   return (
